% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findwb-package.R
\docType{package}
\name{findwb-package}
\alias{findwb}
\alias{findwb-package}
\title{findwb: Download and Clean World Bank Data}
\description{
To learn more about findwb, see the individual help files for each dataset.
}
\author{
\strong{Maintainer}: Anna Mantsoki \email{anna.mantsoki@finddx.org}

Authors:
\itemize{
  \item Mike Page \email{mike@cynkra.com}
  \item Andrea Ranzato \email{andrea@cynkra.com}
}

Other contributors:
\itemize{
  \item FIND [copyright holder, funder]
}

}
\keyword{internal}
