% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wb_gdp}
\alias{wb_gdp}
\title{World Bank GDP}
\format{
A data frame with 2,128 rows and 4 columns:
\describe{
\item{country}{Country name}
\item{country_code}{3 letter ISO country codes}
\item{year}{Years, currently from 2015:2022}
\item{gdp}{GDP (current US$)}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
wb_gdp
}
\description{
GDP, sourced from the World Bank API.
}
\keyword{datasets}
