% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wb_pop_density}
\alias{wb_pop_density}
\title{World Bank population density}
\format{
A data frame with 2,128 rows and 4 columns:
\describe{
\item{country}{Country name}
\item{country_code}{3 letter ISO country codes}
\item{year}{Years, currently from 2015:2022}
\item{pop_density}{Population density (people per sq. km of land area)}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
wb_pop_density
}
\description{
Population density, sourced from the World Bank API.
}
\keyword{datasets}
