% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wb_pop_total}
\alias{wb_pop_total}
\title{World Bank population total}
\format{
A data frame with 2,128 rows and 4 columns:
\describe{
\item{country}{Country name}
\item{country_code}{3 letter ISO country codes}
\item{year}{Years, currently from 2015:2022}
\item{pop_total}{Population, total}
}
}
\source{
\url{https://datatopics.worldbank.org/world-development-indicators/}
}
\usage{
wb_pop_total
}
\description{
Population total, sourced from the World Bank API.
}
\keyword{datasets}
