% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wb_pop_urban}
\alias{wb_pop_urban}
\title{World Bank urban population}
\format{
A data frame with 2,128 rows and 4 columns:
\describe{
\item{country}{Country name}
\item{country_code}{3 letter ISO country codes}
\item{year}{Years, currently from 2015:2022}
\item{pop_urban}{Urban population (\% of total population)}
}
}
\source{
\url{https://datatopics.worldbank.org/world-development-indicators/}
}
\usage{
wb_pop_urban
}
\description{
Urban population, sourced from the World Bank API.
}
\keyword{datasets}
